generic_ocr_service/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
httpx 
pydantic-settings # Thêm thư viện này